Crate nautilus_persistence

NautilusTrader is an open-source, high-performance, production-grade algorithmic trading platform, providing quantitative traders with the ability to backtest portfolios of automated trading strategies on historical data with an event-driven engine, and also deploy those same strategies live, with no code changes.

Feature flags

This crate provides feature flags to control source code inclusion during compilation, depending on the intended use case, i.e. whether to provide Python bindings for the main nautilus_trader Python package, or as part of a Rust only build.

ffi: Enables the C foreign function interface (FFI) from cbindgen

python: Enables Python bindings from pyo3

Modules

arrow

Defines the Apache Arrow schema for Nautilus types.

backend

Provides an Apache Parquet backend powered by DataFusion.

python

Python bindings from pyo3.