Struct nautilus_persistence::backend::session::DataBackendSession

pub struct DataBackendSession {
    pub chunk_size: usize,
    /* private fields */
}

Provides a DataFusion session and registers DataFusion queries.

The session is used to register data sources and make queries on them. A query returns a Chunk of Arrow records. It is decoded and converted into a Vec of data by types that implement [DecodeFromRecordBatch].

Fields
chunk_size: usize
Implementations

impl DataBackendSession

pub fn new(chunk_size: usize) -> Self

pub fn write_data<T: EncodeToRecordBatch>( data: &[T], metadata: &HashMap<String, String>, stream: &mut dyn WriteStream ) -> Result<(), DataStreamingError>

pub async fn add_file_default_query<T>( &mut self, table_name: &str, file_path: &str ) -> Result<()>where T: DecodeDataFromRecordBatch + Into<Data>,

pub async fn add_file_with_custom_query<T>( &mut self, table_name: &str, file_path: &str, sql_query: &str ) -> Result<()>where T: DecodeDataFromRecordBatch + Into<Data>,

pub async fn get_query_result(&mut self) -> QueryResult<Data>

impl DataBackendSession

pub fn new_session(chunk_size: usize) -> Self

pub fn add_file( slf: PyRefMut<'_, Self>, table_name: &str, file_path: &str, data_type: NautilusDataType )

pub fn add_file_with_query( slf: PyRefMut<'_, Self>, table_name: &str, file_path: &str, sql_query: &str, data_type: NautilusDataType )

pub fn to_query_result(slf: PyRefMut<'_, Self>) -> DataQueryResult